refactor(resolve): moves 'compileResolveOptions' into main/resolveOptions/normalize
because normalization is what was happening there anyway
